Final version

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: vsenniko <vsenniko@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 11:05:42 by vsenniko          #+#    #+#             */
-/*   Updated: 2024/11/01 12:32:51 by vsenniko         ###   ########.fr       */
+/*   Updated: 2024/11/08 15:50:14 by vsenniko         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ static int	ft_isspace(char ch)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long	ft_atoi(const char *str)
 {
 	int		i;
 	int		minus;
